chore: move `flagtag.cpp` to `src/physics/flagtag.cpp`
