+++ src/physics/flagtag.cpp
@@ -1,4 +1,4 @@
-#include "flagtag.h"
+#include "physics/flagtag.h"
 #include "physics/init.h"
 
 constexpr double WHEEL_DIST_SQUARED = 0.64;
